app/requirements.txt
Flask==0.12
requests==2.13.0
ruamel.yaml==0.13.13
redis==2.10.5
